rag-service/requirements.txt
fastapi
uvicorn
transformers
faiss-cpu
pydantic
sentence-transformers